move dataHasher to common
